+++ src/Faker.Tests/InternetTests.cs
@@ -13,7 +13,7 @@ namespace RimuTec.Faker.Tests {
 
          // assert
          AllAssertions(domainName);
-         Assert.AreEqual(1, Regex.Matches(domainName, @"\.").Count);
+         Assert.AreEqual(1, RegexMatchesCount(domainName, @"\."));
       }
 
       [Test]
@@ -47,7 +47,7 @@ namespace RimuTec.Faker.Tests {
 
          // assert
          AllAssertions(emailAddress);
-         Assert.AreEqual(1, Regex.Matches(emailAddress, @"@").Count);
+         Assert.AreEqual(1, RegexMatchesCount(emailAddress, @"@"));
       }
 
       [Test]
@@ -82,15 +82,62 @@ namespace RimuTec.Faker.Tests {
          var userName = Internet.UserName("Nancy");
 
          // assert
+         AllAssertions(userName);
          Assert.AreEqual("nancy", userName);
+      }
+
+      [Test]
+      public void UserName_With_Given_Separators() {
+         // arrange
+         char[] separators = new[] { '-', '_', '+' };
+
+         // act
+         var userName = Internet.UserName(separators: separators);
+
+         // assert
+         AllAssertions(userName);
+         Assert.GreaterOrEqual(1, RegexMatchesCount(userName, @"[-_+]"));
+      }
+
+      [Test]
+      public void UserName_With_No_Separator_Guaranteed() {
+         // arrange
+
+         // act
+         var userName = Internet.UserName(null, new char[0]);
+
+         // assert
+         AllAssertions(userName);
+         Assert.AreEqual(0, RegexMatchesCount(userName, @"[-_+]"));
+      }
+
+      [Test]
+      public void UserName_With_String_Argument() {
+         // arrange
+         var firstName = "bo";
+         var lastName = "peep";
+
+         var tries = 10;
+         while(tries-- > 0) {
+            // act
+            var userName = Internet.UserName($"{firstName} {lastName}");
+
+            // assert
+            AllAssertions(userName);
+            Assert.IsTrue(Regex.Match(userName, @"(bo(_|\.)peep|peep(_|\.)bo)").Success, $"userName was '{userName}'");
+         }
       }
 
       private static void AllAssertions(string candidate) {
          Assert.IsFalse(string.IsNullOrWhiteSpace(candidate));
          Assert.IsFalse(candidate.Contains("#"));
          Assert.IsFalse(candidate.Contains("?"));
-         Assert.AreEqual(0, Regex.Matches(candidate, @"[\s]").Count);
-         Assert.AreEqual(0, Regex.Matches(candidate, @"[A-Z]").Count);
+         Assert.AreEqual(0, Regex.Matches(candidate, @"[\s]").Count, $"Candidate was {candidate}");
+         Assert.AreEqual(0, Regex.Matches(candidate, @"[A-Z]").Count, $"Candidate was {candidate}");
+      }
+
+      protected static int RegexMatchesCount(string input, string pattern) {
+         return Regex.Matches(input, pattern, RegexOptions.Compiled).Count;
       }
    }
 }