+++ releasenotes.md
@@ -3,7 +3,7 @@
 ## Version 0.18.0
 ### New Features
 - Internet.Email(string = null)
-- Internet.UserName(string = null)
+- Internet.UserName(string name = null, char[] separators = null)
 
 ## Version 0.17.0
 ### New Features